Create sourceCodeEnvironment
Summary:
I want to make it possible to turn UnannotatedGlobalEnvironment
into a normal environment functor instance (using `environment.ml`).

To do this, we need a PreviousEnvironment with the right interface,
which means that `Environment.PreviousEnvironment` cannot depend on
UnannotatedGlobalEnvironment at all; the prececing diff fixed most of
the interface but creation still required an `UnannotatedGlobalEnvironment.CreateHandle`.

My solution here is simple: make a new environment layer whose interface
is *exactly* what is required by `Environment.PreviousEnvironment`
and whose implementation type is exactly the same record that was
previously `UnannotatedGlobalEnvironment.CreateHandle`.

This new logic has to be implemented in a separate ocaml file from
SourceCodeIncrementalApi because we do *not* want SourceCodeIncrementalApi
to be aware of `AstEnvironment` (this is what guarantees that per target
type checking can use a completely separate implementation), whereas
we *do* require that the base environment layer be ware that it *might*
have access to an `AstEnvironment`; as a result we need a separate module
to avoid cycles.

To keep diffs clean and easy I just implement the new environment in this
diff, I'll cut over to it in the next.

Reviewed By: grievejia

Differential Revision: D58853168

fbshipit-source-id: d7b2bb6343a851e416ea3cf043296c555a8ac90c