Create functionDefinitionEnvironment
Summary:
Pyre saw only a very mild perf regression in D58842854, but Pysa
saw a large bump in RAM use and a big wall time regression (which appears
to be caused by pushing into swap).

The underlying problem is probably that storing the function definitions
in `Module.Components` and bulk-loading it from shared memory into a
process-local cache is expensive - every read of UGE causes the entire
AST to be pulled into process-local memory, and we eventually duplicate
many of the module ASTs in *every* worker.

The underlying problem is really that the ideal data model for function
defnitions, which are rarely-loaded and very big, is just incompatible with the
ideal data model for the rest of `Module.Components`, which is small and
heavily used.

If we assume preprocessing is fairly cheap, this is actually not all that
hard to solve because we can just separate them into two different environment
layers: UnannotatedGlobalEnvironment will handle all of the symbol table data
in a dependency tracked table with cache, whereas functionDefinitionEnvironment
will do the same but in a dependency tracked table with no cache.

**Perf impact**

*Abstractly what's expected*

We're mainly expecting reductions in RAM use here, although there's
also a decrease in serialization since, when loading signature
information (e.g. a class definition) we no longer have to load the
entire AST.

*Effect on Pyre*

This diff actually produces a small but stat-sig *improvement* in Pyre
wall times (-3.87%) relative to trunk (versus the previous diff where we saw a
small and not stat-sig regression of about 3%).

*Effect on Pysa*

This seems to help a bit, every phase of Pysa analysis is at least a
few minutes faster than on the previous diff.

But callgraph construction is still quite slow; 55m on the diff vs
13m on base. So while it helped a bit, it didn't have a huge impact.

**Semantic Changes**

After talking with Jia, I think this change *might* be good in the long run for not just
perf but actually semantics as well. The reason is that it allows UGE and function
definitions to diverge over time; we might be able to make the UGE eventually use
non-preprocessed ASTs, delaying wildcard expansion and the bits of preprocessing
we do need to be "just in time" in later analysis phases.

This is still a vague and aspirational idea, but the fact that we even had a serious
conversation about it makes me think separating UGE (the module *signature*
extraction) from FunctionDefinition data is a pretty good idea for more than just
perf and data layout reasons.

**Diff notes**

The Pysa model changes look scary in the summary, but they are actually trivial;
with this change I no longer make it impossible to get at the function definitions of
external code, and as a result we are able to find the locations of the builtin source
and sync; in real pysa runs all code is marked as in-project anyway so this change
to the unit tests is irrelevant to actual pysa runs.

The locationBasedLookup changes are similar; we're able to resolve slightly
more locations and docstrings after this change because it's now possible to fetch
the ast of a dependency if you really want to.

All the other changes are just mechanical consequences of removing
`function_definitions` from `Module.Components.t` and injecting a new environment
layer in between `AnnotatedGlobalEnvironment` and `AttributeResolution`.

Some test cases disappear but I actually don't think we really need them; the old
UGE tests were excessive because it was a special hand-rolled environment that didn't
use the normal mostly-foolproof functor, but at this stage the dependency tracking is pretty
foolproof and I don't think it actually needs a bunch of update unit tests, the integration tests
should suffice. I could add the tests back as a follow-up if anyone feels strongly.

Reviewed By: migeed-z

Differential Revision: D58897396

fbshipit-source-id: 31b46bfa4dbc2b5b48c04c6b8c4b61cae63c8790