Make UGE a normal environment
Summary:
This commit is a big and potentially signifincantly perf-impacting
change. It gets rid of the complex many-table layout of UGE and makes UGE
a table of a single `Module.Components` value per python module.

**Some details**

The UnannotatedGlobalEnvironment changes are actually pretty
self-explanatory: it now uses a normal environment table keyed
on the qualifier with all the components; the new code follows
in a pretty obvious way and it is now *vastly* simpler.

The tests for UGE are mostly unchanged, but because dependency
tracking happens at the table level, we don't prune quite
as many
dependencies as before in the update tests.

The less-obvious changes are in AstEnvironment:
- we no longer filter dependencies, this is directly implied by
  relying on the normal environment functor in UGE.
- we now *do* filter overlay update `invalidated_modules`,
   because UGE can't since it is functor-based. This is a
good change
  anyway; AstEnvironment ought to handle it's own update results.

**Performance impact**

*Abstractly what is expected here?*

This has a pretty big impact on RAM use because now a worker that
loads any symbol from a module winds up loading all the data instead
of just what is used. It is also expected to affect CPU a bit because of
more deserialization, although that effect is harder to predict since
loading more data but in bigger chunks could actually be more efficient.

*Actual impact on Pyre*

The impact on Pyre is small; it looks like we slow down about 3%, and
the loss isn't even quite stat sig. I'm pretty confident we'd be willing
to take this hit, although in the stacked diff D58897396 I optimize
things a bit and we actually get a stat-sig almost 4% *win*.

*Actual impact on Pysa*

Pysa is much more affected - the taint analysis time is almost
unchanged, but the callgraph construction is quite a lot slower. We're
pretty sure from cross-referencing the sandcastle memory usage
plots that we're pushing well into swap almost immediately, which
causes everything to slow to a crawl.

The overall slowdown is very substantial here, total analysis almost
doubled from 57m to 1h 54m.

I'll stack some diffs with a few mitigations; they do seem to help
some although I'm not sure there's a way to avoid any impacts on
Pysa.

Reviewed By: grievejia

Differential Revision: D58842854

fbshipit-source-id: 3579b59d942e2d0745f5b1b22f96d04fdb14729c