Remove never-ran pyre-codenav test
Summary:
This test was added in D41615576, but never added to the list of tests to
actually run, so it never ran. This was exposed when upgrading the dune lang
version (later in this stack).

This diff just removes the test. I tried fixing it, but the test assertions have
become stale and it didn't seem worth it.

Reviewed By: grievejia

Differential Revision: D59380842

fbshipit-source-id: 3f722ccb4ce5ce7ced3fd17e55533ac771a934f8