Require 2 type params for Annotated
Summary:
This fixes a conformance issues in the `qualifiers_annotated` test, and adds the requirement that `Annotated` must be subscripted with at least 2 arguments.

Since our internal representation of the type was removed in D36736327 and it gets automatically unwrapped during parsing now, I had to sort of hack this in.

Now, we only unwrap it if there are >=2 type params; otherwise we leave it wrapped and check it like a variadic generic that requires >=2 params.

Reviewed By: stroxler

Differential Revision: D59278090

fbshipit-source-id: 28ff6f9b970e2c855db6c2bc80aea6f02bccdbbe