Change `variable` to `name` in `Type.Variable.TypeVar.record`
Summary:
The name `variable` was weird, it led to a lot of code like
`let name = variable.variable` or `match TypeVar { variable = name; _ }`,
and it is inconsistent with all the other type-var-like types, whose
names are called `name`.

I decided to fix this now because I'm about to make a new type-var representation
where the `bound` isn't resolved yet, and I didn't want to duplicate a poorly
chosen record name.

Reviewed By: yangdanny97

Differential Revision: D59497224

fbshipit-source-id: a4412f861538aeead267520103654d7ea8609c4a