create should take variables and aliases as parameters
Summary:
Context:

We are trying to separate processing variable aliases and type aliases in preparation for PEP695.  This entails getting rid of `Alias.t`

As such, the `create` would not be dealing with variable aliasing itself so I am passing variable aliases as parameter, together with aliases which is also another parameter.

More diffs coming soon with the goal of getting rid of `Alias.t`

Reviewed By: stroxler

Differential Revision: D59384363

fbshipit-source-id: e558e3723809ec92231c199af274f5e7a1614ec4