Do not rely on Type.create in parse_instance_annotation
Summary:
The `parse_instance_annotation` in `Type.Variable.ParamSpec`
is able to take annotations like `*args: P.args, **kwargs: P.kwargs`
showing up in a function definition and detect that `P` might be
a ParamSpec (technically we don't know this just from the syntax,
but it takes a callback to look up `P` in the environment whenever
the form matches).

Currently, it depends on `Type.create` because it tries to *convert*
the name of `P` into a type, just so it can extract the primitive name.

This is a problem for me because it causes a dependency cycle I want to
avoid (I really don't want this function depending on `Type.create`, which
is a huge function and introduces an additional unnecessary dependency
on type aliases).

But it's also kind of silly - only a primitive name is valid here, and we
can match that syntactically without anything so heavyweight. This commit
just modifies the code to do a simple syntactical match, and eliminates
the need for `create_type` to be a callback.

Reviewed By: migeed-z

Differential Revision: D59526885

fbshipit-source-id: eb5dccbbf606e92c7aab572aa0a7e9b4d8cf3791