Rename some param spec extraction logic
Summary:
I have always found the name
`parse_as_parameter_specification_instance_annotation`
extremely verbose and at the same time not all that helpful - I actually
had a hard time until now even figuring out what it does, it has a lot of
very long parameter names that weren't obvious to me.

This diff is a pure rename, we now call this function
`param_spec_from_vararg_annotation` and we name the inputs
`~args_annotation` and `~kwargs_annotation`.

I think from this naming scheme it's maybe actually kind of obvious
what it does: given vararg annotations of the form `P.args` and `P.kwargs`
for the `*positional_varargs` and `**keyword_varargs` in a function call,
it extracts the underlying `P` as a `Type.Variable.ParamSpec.t`.

The names `args` and `kwargs` for the positional varargs and keyword_varargs
are technically just convention, which I think maybe explains the verbose
(and inconsistent!) names the code was using for them, but that idiom is
so pervasive that it is cannonical, and I think we should rely on it in
our naming convention rather than making up some less-familiar and more
verbose name for the same thing.

Reviewed By: migeed-z

Differential Revision: D59526888

fbshipit-source-id: 83f3541e9348b82078e9ae1e2bcd00c6a4df35dd