Don't pass `aliases` to `parse_instance_annotation`
Summary:
This commit follows up on the previous change by breaking any
connection from `aliases` to `parse_instance_annotation`; we don't
want this connection because in reality a param spec cannot depend
on a type alias at all, and moreover we want to move `Type.Alias.t`
into `TypeAliasEnvironment` where it can be an implementation detail;
we have to break this dependency before we can do that.

The fix is super simple: instead of taking the aliases callback that
exposes type aliases, take a callback that *only* handles param specs
and let `TypeAliasEnvironment` do the work of constructing that instead
of inlining it into the `type.ml` code.

Reviewed By: migeed-z

Differential Revision: D59526887

fbshipit-source-id: fc47db9214b2e359b024ccbc0554743f03d57697