Move `Type.Alias` down
Summary:
There is still some code that currently depends on Type.Alias so I cannot
eliminate it (Zeina is working on this; eventually we want to make it
an implementation detail of TypeAliasEnvironment and never rely on it so
that we no longer assume TypeAlias and type variables are handled in the
same place).

But what I *can* do is move it down, because none of the "core" type variant
submodules depends on it anymore after the changes made in preceding diffs
to `Type.Variable.ParamSpec`.

This is important because moving the alias definition to live underneath
all of the core type submodules will enable me to define a new module
`Type.Variable.Declaration` and make it so that the variables that we
store in the environment use that type instead of `Type.Variable.t`
(this is needed for 695 because we need a way to represent "expression-level
variables" where we haven't yet resolved aliases and converted to `Type.t`;
I'll have more details on why in the diffs where I do this, but for this
commit the important thing is just that I need to move `Alias` lower
down first).

Reviewed By: grievejia

Differential Revision: D59534689

fbshipit-source-id: 0bcf70692dc3dc18397c3cd296eac00cbf7e7a5a