Add triggering sources as a new taint transform
Summary:
# Problem

For context, we currently don't propagate partial sinks, which causes
false negatives.
We could enable propagating them, but the difficulty lies in showing the sink
traces in the UI. One of the core problems is the sink traces are
"disconnected" due to using triggered partial sinks and partial sinks (which
are different sink kinds) to represent the same sink kinds -- while the users
declare partial sinks, partial sinks are turned into triggered partial sinks
when finding a flow.

# Solution 1 (This diff)

One solution is to implement triggered partial sinks as taint transforms.  The
insight is that, taint transforms adds information to taint kinds, which is
exactly what triggered partial sinks were designed to achieve -- we want to
distinguish partial sink kinds that have a flow from those do not.

Implementation-wise, when creating triggered partial sinks, we want to
associate partial sink kinds with the source kinds that result in creating
them, which are called "triggering sources". Hence, we treat triggering sources
as taint transforms.

The benefit is simple implementation when compared with Solution 2. We can
reuse the implementation for taint transforms.

The downside is that it might be error-prone to treat triggered partial sinks.
While it is legal / makes sense to chain a list of taint transforms, such
behavior is undefined for triggered partial sinks (e.g., it makes no sense to
trigger a partial sink twice), but is not yet prevented from happening.

# Solution 2

Another solution is to ensure the sink traces always use partial sinks, and do
not use triggered partial sinks. Then, the triggering sources need to be stored
into `Frame`, as opposed to being stored into the partial sink kinds (which is
the current implementation) -- for context, `LocalTaint` is a map from taint
kinds to `Frame`s.

The downsides are
- less precise `Frame` (e.g., due to merging `Frame`s that used to be
  associated with different sink kinds -- partial sinks vs. triggered partial
  sinks, but now we no longer distinguish them), and
- more complex implementation (e.g., to decide whether to create a flow in
  `issue.ml` that ends with a partial sink, now we need to look into `Frame` to
  see whether the partial sink has been triggered).

These downsides do not exist in Solution 1.

Reviewed By: arthaud

Differential Revision: D59400049

fbshipit-source-id: ef5aa7e5187a0c76f03ab52271e8130b6cd189f4