Add failing test for resolving generator target in condition expressions
Summary:
When a generator's expression contains a reference to the generator target, the
type is not includedin the types() query. Note that the existing code only
handles the case where the condition is exactly a reference to the target, but
not cases where the reference is recursively contained within the condition
expression.

Reviewed By: vladima

Differential Revision: D59504875

fbshipit-source-id: 64ec240fb8ece23b3a29107234a14557289e399c