Re-use existing helper to convert generator into synthetic assignment
Summary:
The code to compute coverage for generators needs to compute a resolution value
including the target. We do this by creating a synthetic assignment AST and then
using some existing code to process that assignment.

While looking at this code, I noticed that it was duplicating an existing helper
which is used for this purpose in typeCheck.ml. Furthermore, the existing helper
handles async generators which this code does not.

This isn't designed to fix anything in particular, although it might fix some
async generator handling. Just submitting this as part of the stack because I
noticed this while reading code.

Reviewed By: vladima

Differential Revision: D59504876

fbshipit-source-id: c924e572555a1db657f4ffb5e52ea1e65475cca1