Use `non_sanitize_transforms` when looking for entries in `matching_source_sanitize_transforms` and `matching_sink_sanitize_transforms`
Summary:
The reason is that, when creating maps `matching_source_sanitize_transforms`
and `matching_sink_sanitize_transforms`, we currently allow transform
`TriggeredPartialSink` to be used in the maps' keys. Hence, when querying these
maps, we should allow constructing keys with transform `TriggeredPartialSink`.

This change does not affect any analysis results (from the last diff), because
querying the maps is needed only when applying / adding taint transforms, but
it is currently disallowed to adding transform `TriggeredPartialSink`.

Reviewed By: arthaud

Differential Revision: D59610211

fbshipit-source-id: d86d1d21ea2733905cf31803db4246a30cb200b9