add local search to site
Summary:
This diff adds local search functionality to the Pyre staticdocs site, improving documentation discoverability. The default search functionality built into staticdocs is slow, not prominently placed, and works poorly, so I think this would be a big upgrade for our users.

How it works:

A search index is generated when the site is built, and downloaded by the client when the page loads. The search happens locally in the browser, so it's lightning fast and gives results in a nice dropdown.

In the future, we can log search queries to get insights on what users are struggling with. that would require swizzling (https://docusaurus.io/docs/swizzling) the search component to make modifications, instead of using it out of the box.

There's also probably some optimizations we can make to improve the performance, like saving the index to the browser's local storage & using the stale one while the new index is loading (bunnylol smaxdocs does this), but let's wait and see how this performs first before prematurely optimizing.

Reviewed By: grievejia

Differential Revision: D59635074

fbshipit-source-id: bdd82990c3f0511fc83ce5de2cae726eef4ebaec