Add Type.Variable.Declaration [2: use it in Type.Alias.t]
Summary:
This commit builds on the new `Type.Variable.Declaration` module
added in the previous diff (see that diff summary for a detailed
explanation of why we want this new concept for PEP 695), making
`Type.Alias.t` use a declaration instead of a fully alias-resolved
`Type.Variable.t`.

This allows all alias handling to be dealt with in `create_logic`,
the declaration parsing *never* does anything more than syntactic
analysis. By reducing type var declaration handling to syntactic
only, we make it much easier for a future change to store scoped type
variables in the style of 695 (which aren't global and can't live in
TypeAliasEnvironment easily) in UnannotatedGlobalEnvironment.

NOTE: the production code part of this diff is quite small; most of
the changes are needed because test code couples to the exact types
inside of `Type.Alias.t` (which is something we want to get rid of
as part of prep for 695 - all these tests eventually should be defining
variables directly, not wrapping them up in the `Alias.t` variant).

Reviewed By: migeed-z

Differential Revision: D59534687

fbshipit-source-id: 07b12513abac1f7dd3028081bba8440e35f760e3