Fix pysa step in github CI after dune lang bump (#896)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/896

Prior to this commit, we were relying on an `apt` installed `dune`
rather than the `dune` from the OPAM switch to build Pyre in the
`pysa` github job.

This wasn't obvious - for reasons I don't really understand right
now, the `env` passed from `setup.py` that includes the opam switch
bin directory on `PATH` isn't making into the `Makefile` action (see
the TODO).

A workaround is to just set the `PATH` manually.

Differential Revision: D59693879