Hide implementation details for type aliasing and variable from parts of the typechecker
Summary:
Context:

We want to hide implementation details for type aliasing and variable aliasing. The reason is that we need to get to a point where we can completely hide those implementation details from the typechecker and this diff is a step towards that.

Changes:

We grab the type aliasing and variable aliasing functionality from the initial state and pass it to various parts of the typechecker. There are other places that still know about the implementation details. I willl consider those in upcomming diffs.

Reviewed By: stroxler

Differential Revision: D59604780

fbshipit-source-id: bac43c1aa51383ce5b2d3b7f74a2b34c46db4439