Remove unused codenav error messages which were duplicates of empty reasons
Summary: D55893688 Introduced duplicated logging by including empty reasons as part of error messages in addition to the existing empty reasons field, since we needed to wait for devs to rebase on top of D55731548. It has been many months since then so that this is good to clean up.

Reviewed By: kinto0

Differential Revision: D59637831

fbshipit-source-id: ae5faf61ed45b29bb06716e454e8870c20add2e6