Add variable and type alias functionality in Resolution
Summary:
Following up on the context from this diff stack, we are trying to eliminate all uses of Alias.t from the typechecker.  We will eventually move Alias.t to the type alias environment and let it be an implementation detail that is not exposed to the typechecker.

This diff removes all references of these implementation details from the typechecker.

We will followup with a cleanup for the callbacks from the typechecker in this diff stack.

Reviewed By: stroxler

Differential Revision: D59652473

fbshipit-source-id: ce8c4567accd807e9a878d9b9e1b858b4a7e7dca