Reorganize TypeGuard code a bit
Summary:
I'm about to add support for `TypeIs` to Pyre, and a majority of the code
should be shared with `TypeGuard`.

There are a few places where the existing code isn't optimized for extending
the concept of type guards to be more general, either due to poor factoring
or names that aren't as clear as they could be. This is a simple commit
with no functional changes to clean some of these up in preparation for
`TypeIs` support.

Reviewed By: grievejia

Differential Revision: D59734070

fbshipit-source-id: c892c2b526fbbc5a1fc17aeb4f74d138c065e316