Ensure that the aliases argument which is passed to Type.create returns a type
Summary:
Context:

We are separating the computation path between type aliasing and variable aliasing in an effort to support PEP695. In this diff, we focus on type.ml where we make sure that the aliases argument to Type.create returns a type.

In upcomming diffs, our goal will be to get rid of Alias.t, which is a variant in type level that represents combining aliases and variables (which is something we don't want).

Reviewed By: stroxler

Differential Revision: D59736330

fbshipit-source-id: 2c2141fd5292f21c9e4edf9b11051c9c36c25512