Add consistency check to type guards
Summary:
The typing spec requires that as a sanity check, a type checker
should verify that a type guard's narrowed type is a (gradual)
subtype of the guarded type, i.e. that the parameter annotation
and the type inside the guard are not entirely inconsistent.

This in particular can warn users on *some* unsound type guards,
for example it is never really safe to narrow the type inside a
list and this check will catch it because it leds to an invariance
error.

Reviewed By: yangdanny97

Differential Revision: D59761173

fbshipit-source-id: 4d9fbed5f5e0aee4e67c55ab15f1cb8cfad339ac