Refactor to prepare for consistency check on typeguards
Summary:
The typing spec requires that type checkers sanity check a type guard by
detecting when the narrowed type is not consistent with the type of
the narrowed parameter.

This is easier to add to the code if we first do a pure (no functional
change) refactor so that the existing logic is designed around extracting
the parameter and return type instead of hand-coded match statements
and parameter list length checks; that way, the existing logic already
has all the data we need to insert the new validation.

This commit does that refactor:
- when constructing `parameter_types` at the top of the outer function,
  actually get the type information (prior to this commit we were just inserting
  `Any`, which led to a lot of confusion for me!)
- rely on `Type.inner_type_of_typeguard_or_typeis` rather than matching the
  name directly (for now throwing away the actual inner type but we'll use it
  in the next commit)
- use `List.nth` to extract the correct positional parameter rather than using
  length checks (for now throwing away the extracted parameter but we'll use it
  in the next commit).

Reviewed By: yangdanny97

Differential Revision: D59761171

fbshipit-source-id: 47e59593d1b84faec10992dde11474940ae2f5ac