Support TypeIs (for the most part)
Summary:
The support is entirely via `missingFromStubs`, which is
actually not as crazy as it sounds - the type stubs for special
forms have no meaning whatsoever in Pyre, because Pyre has
a hard requirement that all types are classes and the special
forms are not classes.

*Potentially Controversial* Since `TypeIs` is expected to become the "primary"
type guard used most of the time, I decided to cut over most of the test suite,
leaving only the initial set of tests duplicated; the code for handling
everything except negative-case narrowing is identical so it's not clear that
we gain much by duplicating every test; if the code diverges more in the future
(unlikely??) this could open the door to regressions. If there are any objections
I can go ahead and duplicate more of the tests.

*Bug* I discovered a bug in our negative case type narrowing; we try to do
a reachbility analysis which is itself debateable; in part of that logic we
attempt bypass an "unreachable" conclusion when dealing with gradual types
but we aren't handling all the gradual forms (the example I found is
`Tuple[X, ...]` but there are others) properly.

I made a task for this, I won't fix it yet because we should have a team
discussion first - it might be good to improve the logic, or it might be
better to stop doing reachability in the first place (we've long had
questions about whether this is good to do in the first place).

**Conformance**

It is required that a type checker should catch incompatibility between
the input and narrowed type in a type guard, and we currently don't do this
for `TypeIs` *or* for `TypeGuard`.

The tests only actually verify this for `TypeIs` though; as a result we currently
pass the suite for `TypeGuard` even though support is the same, and we need to
implement the validation before we will pass the `TypeIs` test. I'll work on the
missing validation in a stacked diff.

Reviewed By: yangdanny97

Differential Revision: D59734069

fbshipit-source-id: 3578efd8b92d77abae5a8756dcab21e0cce349d8