Cut most TypeGuard tests to Sequence
Summary:
I'm about to fix a conformance gap in Pyre where we were failing
to check that the narrowed type is a subtype of the guarded type
(this is required for `TypeIs`, not for `TypeGuard`, but I want most
tests to run on `TypeIs` which is now the "primary" type guard used
by typeshed and therefore more important for us).

Once we check this, almost a majority of our unit tests are actually
illegal type guards, because a ton of them are based on narrowing
a `List` type which is always illegal due to invariance.

To keep the impact of the change on tests manageable, I am cutting
most of the tests over to using `Sequence` which is the type for
which most of these checks are valid.

**A note on soundness**

I'd like to note that even with `Sequence` the type narrowing is
actually unsound. Mutable values are awful for sound type refinement;
even a `Sequence` could potentially have live references to it with
read-write types (e.g. lists) such that the narowed type can be
invalidated by mutations happening far away in the codebase. But the
spec does allow it, so pragmatically we should allow this source
of unsoundness at least for now.

Reviewed By: yangdanny97

Differential Revision: D59761174

fbshipit-source-id: af3e344132ca524618072ee2d025a15c65655e07