Clean up type guard code a bit
Summary:
Suggested in review of the stack, and I remember noticing it as
I was working on earlier diffs and deciding to put it off... let's
always use `Type.narrowing_of_type_guard` as the source of truth
for whether a type is a type guard.

But also, after further thought I think the names I chose are
not great; it sounds like `narrowing_of_type_guard` is always
going to be called on a type guard and returns the narrowing,
whereas it can be called on *any* type and we often use it to
ask "is this thing a type guard?".

So, renaming it and also using the CPython names for `TypeGuard`
and `TypeIs` instead of the made-up and more theory-inspired
names `Exact` / `Positive`. I feel like the new names will be
much more self-descriptive for future Pyre devs referencing
the spec and the relevant Python code snippets.

Reviewed By: yangdanny97

Differential Revision: D59842085

fbshipit-source-id: 028dbdd0934cdfb15805607e18625438fed10a20