fix bug in non-modifying decorator allowlist
Summary:
The idea of this piece of logic is that functions whose signatures are potentially modified by decorators are exempt from certain override-related checks. It was added in D45058619.

Certain non-modifying decorators are hardcoded/ignored, and the `override` decorator does not modify the signature, but the hardcoded list needs to use the qualified name for it to work.

There's an existing comment/task T150016653 for replacing this with something better, but I'm just doing this small fix to unblock my work on strict override checks.

Reviewed By: migeed-z

Differential Revision: D59807468

fbshipit-source-id: 3e0ec40d28b5a3e38a64869ccf9c6d684d059884